+++ src/ska_src_site_capabilities_api/rest/static/js/edit-node-form-ui.js
@@ -466,6 +466,10 @@
                                   }
                                 },
                                 {
+                                  "key": "sites[].storages[].areas[].environments",
+                                  "type": "checkboxes"
+                                },
+                                {
                                   "key": "sites[].storages[].areas[].other_attributes",
                                   "type": "ace",
                                   "aceMode": "json",
@@ -631,6 +635,10 @@
                                   }
                                 },
                                 {
+                                  "key": "sites[].compute[].associated_local_services[].environments",
+                                  "type": "checkboxes"
+                                },
+                                {
                                   "key": "sites[].compute[].associated_local_services[].other_attributes",
                                   "type": "ace",
                                   "aceMode": "json",
@@ -715,6 +723,10 @@
                                       }
                                     ]
                                   }
+                                },
+                                {
+                                  "key": "sites[].compute[].associated_global_services[].environments",
+                                  "type": "checkboxes"
                                 },
                                 {
                                   "key": "sites[].compute[].associated_global_services[].other_attributes",