+++ tests/unit/test_backend.py
@@ -10,14 +10,14 @@ from ska_src_site_capabilities_api.backend.mongo import MongoBackend
 @pytest.fixture(scope="module")
 def dummy_nodes():
     """Fixture to return nodes json."""
-    with Path("tests/assets/nodes.json").open("r") as nodes_file:
+    with Path("tests/assets/unit/nodes.json").open("r") as nodes_file:
         return json.load(nodes_file)
 
 
 @pytest.fixture(scope="module")
 def dummy_nodes_archived():
     """Fixture to return nodes_archived json."""
-    with Path("tests/assets/nodes.json").open("r") as nodes_file:
+    with Path("tests/assets/unit/nodes.json").open("r") as nodes_file:
         return json.load(nodes_file)
 
 
@@ -158,18 +158,28 @@ def test_list_nodes(mock_backend):
 
 
 @pytest.mark.unit
-def test_list_services_with_output_prometheus(mock_backend):
-    services = mock_backend.list_services(for_prometheus=True)
-    for service in services:
-        assert "targets" in service
-        assert "labels" in service
-        assert service["targets"]
+@pytest.mark.parametrize(
+    "environment, number_of_services",
+    [("Production", 3), ("Development", 2), ("Integration", 4)],
+)
+def test_list_services_with_environment_filter(mock_backend, environment, number_of_services):
+    services = mock_backend.list_services(environments=[environment])
+    assert len(services) == number_of_services
 
 
 @pytest.mark.unit
 def test_list_services_with_node_name_filter(mock_backend):
     services = mock_backend.list_services(node_names="TEST")
     assert len(services) == 7
+
+
+@pytest.mark.unit
+def test_list_services_with_output_prometheus(mock_backend):
+    services = mock_backend.list_services(for_prometheus=True)
+    for service in services:
+        assert "targets" in service
+        assert "labels" in service
+        assert service["targets"]
 
 
 @pytest.mark.unit
@@ -189,6 +199,16 @@ def test_list_services_with_site_name_filter(mock_backend):
 def test_list_sites(mock_backend):
     sites = mock_backend.list_sites()
     assert len(sites) == 2
+
+
+@pytest.mark.unit
+@pytest.mark.parametrize(
+    "environment, number_of_storage_areas",
+    [("Production", 1), ("Development", 2), ("Integration", 1)],
+)
+def test_list_storage_areas_with_environment_filter(mock_backend, environment, number_of_storage_areas):
+    services = mock_backend.list_storage_areas(environments=[environment])
+    assert len(services) == number_of_storage_areas
 
 
 @pytest.mark.unit