+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.65]
+
+### Added
+
+- Service environments (int, prod, dev)
+- Refactored bits of integration tests
+
 ## [0.3.64]
 
 ### Added